usb: dwc2: gadget: disable ep0 when stop udc
Test usb hotplug on the PX30/RK3308 boards, it may fail
in the enumeration phase. It's difficult to reproduce in
practice for user, but if we do quickly hotplug (more than
3 times per second), it can be reproduced more easily.

The typical error log on the PX30 boards like this:
phy phy-ff2c0000.syscon:usb2-phy@100.1: charger = USB_SDP_CHARGER
dwc2 ff300000.usb: new device is high-speed
dwc2 ff300000.usb: new device is high-speed
dwc2 ff300000.usb: new device is high-speed
...
dwc2 ff300000.usb: new device is high-speed

The root cause is that we don't disable the ep0 and stop
the transfer on ep0 if it's enabled when plug out usb.
So the ep0 fails to transfer when connect next time.

To fix this issue, this patch disable the ep0 in the
udc_stop() ops which can be executed when plug out usb.

Signed-off-by: William Wu <william.wu@rock-chips.com>
Change-Id: I0dd313540ce52a6d01a6223e89c8cd67e6815965